refactor: remove usePopoverRoot
